Move test classes to test main folder; (#2807)
## Summary
This pr does move test files to the test folder of the project.

## Screenshots
NA

## Link to pull request in Documentation repository
-

## Any other notes
-
