+++ Sources/App/Utilities/Utils.swift
@@ -74,16 +74,3 @@ func setDefaults() {
         prefs.setValue(true, forKey: "confirmBeforeOpeningUrl")
     }
 }
-
-extension UIImage {
-    func scaledToSize(_ size: CGSize) -> UIImage {
-        UIGraphicsImageRenderer(
-            size: size,
-            format: with(UIGraphicsImageRendererFormat.preferred()) {
-                $0.opaque = imageRendererFormat.opaque
-            }
-        ).image { _ in
-            draw(in: CGRect(origin: .zero, size: size))
-        }
-    }
-}
